requirements.txt
Flask==3.0.3
fpdf==1.7.2
google-generativeai==0.5.2
